+++ src/main.ts
@@ -1,15 +1,15 @@
-import { ImgBank } from "./loading";
+import { ImageBank } from "./image_bank";
 import { Renderer } from "./renderer";
 import { GameState } from "./gamestate";
-import { RoomMap } from "./room_map";
-import { Settings } from "./settings";
-import { KeyMapper } from "./keymapper";
+import { RoomMap } from "./environment/room_map";
+import { Settings } from "./settings/settings";
+import { KeyMapper } from "./settings/keymapper";
 
 export const canvas = document.querySelector("#canvas") as HTMLCanvasElement;
 export const canvas_W = 640;
 export const canvas_H = 480;
 export const ctx = canvas.getContext("2d");
-export const bank = new ImgBank();
+export const bank = new ImageBank();
 export const renderer = new Renderer();
 export const gameState = new GameState(new RoomMap(0));
